refactor: domain에서 불필요한 클래스 테스트로 이동
